Refactor cart addItem function to accept additional data
The addItem function in the cart actions file has been refactored to accept additional data along with the selected variant ID. This change allows for the inclusion of additional attributes when adding an item to the cart. The selected variant ID and additional data are now passed as an object in the formData parameter. This improves the flexibility and extensibility of the addItem function.

Co-authored-by: theaussie86